Splitting prepare script for convenience
