+++ ios/SalesforceReactTests/updatesdk.js
@@ -1,0 +1,15 @@
+#!/usr/bin/env node
+
+var packageJson = require('./package.json')
+var execSync = require('child_process').execSync;
+var path = require('path');
+var rimraf = require('rimraf');
+
+var sdkDependency = 'SalesforceMobileSDK-iOS';
+var repoUrlWithBranch = packageJson.sdkDependencies[sdkDependency];
+var parts = repoUrlWithBranch.split('#'), repoUrl = parts[0], branch = parts.length > 1 ? parts[1] : 'master';
+var targetDir = path.join('mobile_sdk', sdkDependency);
+rimraf.sync(targetDir);
+execSync('git clone --branch ' + branch + ' --single-branch --depth 1 ' + repoUrl + ' ' + targetDir, {stdio:[0,1,2]});
+rimraf.sync(path.join('mobile_sdk', 'SalesforceMobileSDK-iOS', 'libs', 'SalesforceReact', 'package.json')); // confuses metro bundler
+
